+++ src/statespace/CartesianProduct.cpp
@@ -134,8 +134,7 @@ void CartesianProduct::expMap(const Eigen::VectorXd &_tangent,
   auto dimension = getDimension();
 
   // TODO: Skip these checks in release mode.
-  if ((unsigned int)_tangent.rows() != dimension) // Converting _tangent.rows to unsigned as it cannot be negative anyway
-  {   
+  if (static_cast<size_t>(_tangent.rows()) != dimension) {   
     std::stringstream msg;
     msg << "_tangent has incorrect size: expected " << dimension << ", got "
         << _tangent.rows() << ".\n";
@@ -157,8 +156,7 @@ void CartesianProduct::logMap(const StateSpace::State *_in,
 {
   auto dimension = getDimension();
 
-  if ((unsigned int)_tangent.rows() != dimension) // Converting _tangent.rows to unsigned as it cannot be negative anyway
-  {
+  if (static_cast<size_t>(_tangent.rows()) != dimension) {
     _tangent.resize(dimension);
   }
